+++ lib/config_verify/configVerify_test.go
@@ -21,8 +21,10 @@ func TestConfigValidateOnPrem(t *testing.T) {
 		cv := &ConfigVerifyImpl{}
 		err := cv.ConfigValidateOnPrem(config)
 		if err != nil {
-			t.Errorf("ConfigInitials validation failed: %s", err)
+			return
 		}
+		// If there is no error, fail the test case
+		t.Fail()
 	})
 
 	t.Run("Test Success AWS Config Validation", func(t *testing.T) {
@@ -34,4 +36,37 @@ func TestConfigValidateOnPrem(t *testing.T) {
 			t.Errorf("ConfigInitials validation failed: %s", err)
 		}
 	})
+
+	t.Run("Test Failed AWS Config Validation", func(t *testing.T) {
+		config := getFailedTestAWSConfigData()
+
+		cv := &ConfigVerifyImpl{}
+		err := cv.ConfigValidateAWS(config)
+		if err != nil {
+			return
+		}
+		// If there is no error, fail the test case
+		t.Fail()
+	})
+
+	t.Run("Test Success Automate SC Config Validation", func(t *testing.T) {
+		config := getAutomateScConfigTestData()
+		cv := &ConfigVerifyImpl{}
+		err := cv.ConfigValidateStandalone(config)
+		if err != nil {
+			t.Errorf("ConfigInitials validation failed: %s", err)
+		}
+	})
+
+	t.Run("Test Failed Automate SC Config Validation", func(t *testing.T) {
+		config := getFAiledAutomateScConfigTestData()
+		cv := &ConfigVerifyImpl{}
+		err := cv.ConfigValidateStandalone(config)
+		if err != nil {
+			return
+		}
+		// If there is no error, fail the test case
+		t.Fail()
+	})
+
 }
