+++ lib/config_verify/testdata/admin-key.pem
@@ -1,0 +1,28 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvwIBADANBgkqhkiG9w0BAQEFAASCBKkwggSlAgEAAoIBAQC1UxirTLrVqYE2
+TpZPb8EXtCRWwDXxJn0y8JpaVOU4f9bJJXY9t8Sy+Tiy5gELqFtSqvovWoMLnKtV
+JJNzFvcN+Eup+q63NjIqtaPwF2DWPKcwWwf/DZ6eaXkrZ3eTyT0FMJGgMIiZ9muR
+XBsmOoDrun1JalcFdSKONDjbmRD+14ct1oPJzd9wEF7qdA7xsnSsfCb/R/n6BT3/
+q9Q91UA9SHEZ18Lsmh8YeY1vrr83RrMDn1O5Cxb37Fas2ND953nqMU3UzjJUy9WA
+ymP9MrIrujX2ct6itFpDSxmrj5gKUVE6c92RYAfXrF/3UzOTnUXC25AE0wm45qcF
+squ3pkRZAgMBAAECggEBAIuRm6J3HDTMFT5E/WPFpH5nVbVEnACz3JiyJE01nX6A
+wEXxxPK6mfgc5yyk59ceTrTyEEMUy1vLywFtduWWnrGtFiiPYZ56Y+EGEVsD07Dz
+3MjWuO/0fsx2aSMksCEcPBDbkqwXz32a9K9dkzQ4wfUwhbFV3gfAYsnMnptBVZih
+5o6vgAB90YsLa5+W2hmefNA0gbV2fzNgAFyRkVeFvWTl2c6TA8pEktszcRZG8+Ut
+Av7PCKEM+hW6LOKmf0KM/jH3ZSrfkkD4ixC+SjiKwxoYYwchIAU5iyH/rV4Qo/rx
+AN9KIsxrNYQ7tpp/zhm/Slp2uzqfZwemeC52x23fLSUCgYEA6b2kC6XacJyFdIa0
+T2DXsjsZ5amlybolOwfoogsiYuFvz9JwiC8lhDXDpJ+o0NnH66qHdZ0zvT2TFevH
+1dkG/oRUstFx+d0kgIsCGFd+JQfEx4dZda5pcdXdfdVWU8fmmRTGQlVgwyqRXhqD
+D1T/eCsh426PqUgCTxeENGIpyqcCgYEAxpeaWkzqdyuJDge4NWvgVae+aiV/aY9t
+V4wUdwXBa3NtfyYETcogIPLY19a03rLUO6OFRhOSPFNqGnXZq74yTL7CesbDUfxp
+jkTLt1Nyh5BesrKx/TqfMvgMV7Vegrew+KBkhfPouecRr4W03xeULPXP8bBuT8PK
+u00SKhJEWP8CgYEAnKvYgVYyDLldWrXPtkUbTwP+X8pOreBvhBnQ/IyqB81DvPE5
+LI8tfmjrMjUyWosfTf+HvHYacAYNyOu5jsVhwsfAXfu1gd2Q0fha9yXuD7QdWgCH
+HjJBy2Omiumsa4vaLVGXcm735NlrhsAOiQBYHJwMov1Anh7+LTV+D/6cNJECgYAf
+hTyIjKDCXXSZd3wzbnXi4ocEjgMVHCC8vo0Y/x7eESoV/lC8kBJdaG4mg/udRUEO
+XjvvDXgufwbeQEpWSg/twnFXz2h9ANU4MY8r/+Czx6RgNKCvkrhl/Jzo5u6jTMiY
+Dc0mvbn4l8Zk15IHvtRT9ogU0m4P1wG8jNUPPGsAvQKBgQCJo2X+3cCi2EGYOTwU
+I/EOURoPHMaGqqzwtRYbGKbtYqIK1pJPmYn6Iyea5rw//sSYPDT6OxCWgS17/Ekj
+1vuk4Tgz1W+8sbH14rjK47SAdNFp7yS5fxbZrTrwzqKoI8XHOjlynGAIEuJ0xlkK
+47A9Od3C9GIM8Z8G/npJTXkVpg==
+-----END PRIVATE KEY-----
