+++ lib/config_verify/testdata/admin.crt
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDrTCCApWgAwIBAgIRAKZR+mydckDpDotkgge9GVMwDQYJKoZIhvcNAQELBQAw
+VDELMAkGA1UEBhMCVVMxFjAUBgNVBAoTDUNoZWYgU29mdHdhcmUxFjAUBgNVBAsT
+DUNoZWYgQXV0b21hdGUxFTATBgNVBAMTDGUyYmM1Y2Q2NWVmODAeFw0yMzA1MTUx
+MjA1MjBaFw0zMzA1MTIxMjA1MjBaMFQxCzAJBgNVBAYTAlVTMRYwFAYDVQQKEw1D
+aGVmIFNvZnR3YXJlMRYwFAYDVQQLEw1DaGVmIEF1dG9tYXRlMRUwEwYDVQQDEwxl
+MmJjNWNkNjVlZjgwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDYMeK2
+mKLwHAm/IdNp1ZvPaqQdeZPHXvBBUkNwhisuICTl+NWn6wt+zlEamMnR/CJp58oE
+5MAUE2SRcvhlEHjKJO5O7xGrKp148FQOOCukYbxC990CaCLme1Xks0KygatD3TzT
+/vJ1aMhGRW8MagFx4OYf/2MW/WMGNrVMUzFglg0OUCC35Zu00fV2OEO76Y8CuKWS
+l91cgSERcsiHYLIhEr53ImBCsgvTAUHCCn+GDdIEe65YDhxbboVAliLFK0skDtJB
+1umyZLM1b3szWrA3mFvTAnnXlj8XaSyhB4A9RWdiddxvtajMYuYZOIMev532NIyn
+BIAt80tcsLzPs0dNAgMBAAGjejB4MA4GA1UdDwEB/wQEAwICpDAdBgNVHSUEFjAU
+BggrBgEFBQcDAQYIKwYBBQUHAwIwDwYDVR0TAQH/BAUwAwEB/zAdBgNVHQ4EFgQU
+cW8SxcfLEFgEM3lVPrEgM77ABHYwFwYDVR0RBBAwDoIMZTJiYzVjZDY1ZWY4MA0G
+CSqGSIb3DQEBCwUAA4IBAQDBn+tLyRQ7XWOVbSF6b9WqOUNd26sXZCFD81NicyEU
+wKf6+ZUfxy5XjrXv5JvsIbBlnG6JPHQ/RGCal+p3rTAMVAMzurTuE5Jz+MZTJV5V
+lnh8Cn7kDDTgQOzZ5Ouf/HMC+0IjlaH7Drv+9Hf0gSaqJXD8+pkA9QQGs5O6+NTx
+alVulHFKpEcneE4dAycK5W847RY08EIKjBR6ixmZ4yrghwhZTmVXtznVmJ8V0C9X
+ficIxQmBofUpTYH2g/IRqycwdBu7wGbzKZSNKGxvjQmifs3WKdMYkkcd1Ns8rrW0
+Ib9KmE8pectIvWTdBTtb6ABAfWkWRcytYV+YGj7vJwVI
+-----END CERTIFICATE-----
