+++ lib/config_verify/testdata/admin.pem
@@ -1,0 +1,27 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIEpAIBAAKCAQEA2DHitpii8BwJvyHTadWbz2qkHXmTx17wQVJDcIYrLiAk5fjV
+p+sLfs5RGpjJ0fwiaefKBOTAFBNkkXL4ZRB4yiTuTu8RqyqdePBUDjgrpGG8Qvfd
+Amgi5ntV5LNCsoGrQ9080/7ydWjIRkVvDGoBceDmH/9jFv1jBja1TFMxYJYNDlAg
+t+WbtNH1djhDu+mPArilkpfdXIEhEXLIh2CyIRK+dyJgQrIL0wFBwgp/hg3SBHuu
+WA4cW26FQJYixStLJA7SQdbpsmSzNW97M1qwN5hb0wJ515Y/F2ksoQeAPUVnYnXc
+b7WozGLmGTiDHr+d9jSMpwSALfNLXLC8z7NHTQIDAQABAoIBAQC431/JWRjsTm2f
+PeO1eu50HfMdKVvPf0Xf9eKplHHUTbR/25uIrhv0YaVFIyZiJEOhYI6GRL5j7Hlx
+T2lNsxo3cgE/CnyHbDegtZKLXkkjllgQxDf/2OEPSqSWnxwH3Y6Ax+0+M6NN9z5b
+qQBfEz8NDI59CjzH6gIFzD6YeAOfXN6NqDnm6scLH7Adk2R0qDlnTtcbOXVWh3fc
+XVv56HDSgm0JhqWgfu2bg92sLeHj70oFCsTtyVR4AboSea3cfLsuLs1jBd2VOasd
+AMGdMEMtMD3gl+qofjU8L0RLTlR90+jN0+1n2G+j/9nTJDRO17Zwy7+/tr3If0bp
+RjzSuFWZAoGBAOyAy2606KUNxfdt4oY32R38/Nsz5tcelHKLVHyLhTegwt1ojfbk
+a2dm1C6p9wqz1VzYsSaY5KCkB1ao5sDRHPbQhBTs2gMWM1wT4bDfJD1ACZsOBtnv
+wfPKxUBNd6i8MIuwyC5X9FYREOti2ajX9qnMhunHlgkH4G9ulE+nM2qfAoGBAOoE
+gIpdhdky0LP/a6FJD5Bsy7Qa6wRlf37ctC442O90vXoKIcbhEtc1JKeniBYU5hsg
+BTYBHRpSDB0rUvYVcgOM/T47kNi1k3NLV7F4WvR0QfNsv5FlTu/mpx5aQXcG/G9L
+6xPXbeVAw1b2ll2TCwmqqNZbS5PLKAxW2/ZZZjKTAoGACr8nf1ZVI/rmPgIOQw8S
+nJw8WTGYkbngAPx30Fj+qrJxaGIx4UAr6Tzw81Xh1eiqD1gBZMGZ0friInv2UFiw
+Oj7rYZAJsA44MzzsCn95RBc+2H7grNKKO6vBHtTmVRKIGjDkR8bTYX3RCQEq5tM5
+cCnND550IODsv+OEdcIhQf0CgYEAg6Lur2Z2QTwtW+MEpyAS6jnYyEIQC1QQyMrb
+a9i8Zs4zOy/x2g6Ha5lQOiKqEUnGOmHjrIIp3cFjYREaxKRBky++0D/vdYeGHue1
+t3WtLC8tXgHBLMdhBN+pCBDbHFpIq054OJwJDFlGyGdDI/Xv0os2GAYY2FUB/Rbq
++oNFUpMCgYBEqJym4rr6sbZgy5cdWobJjJAMF0+9H6RBKUuKNgHxnVg/hAYB1Y9B
+v4xVAkmCjUkQhQkYUNa3/L6ybRH6dwWtzXWnUJoMiuhXts0047ywNjL6LrwzP7T2
++a+MEDDwragm3DDT7FcqdVugy54U1nYUyr3rMavmBr6eev3VtQsFzg==
+-----END RSA PRIVATE KEY-----
